+++ src/platform/run_return.rs
@@ -22,7 +22,8 @@ pub trait EventLoopExtRunReturn {
     /// Initializes the `winit` event loop.
     ///
     /// Unlike `run`, this function accepts non-`'static` (i.e. non-`move`) closures and returns
-    /// control flow to the caller when `control_flow` is set to `ControlFlow::Exit`.
+    /// control flow to the caller when `control_flow` is set to `ControlFlow::Exit`, where its
+    /// content representing the exit code is the return value.
     ///
     /// # Caveats
     /// Despite its appearance at first glance, this is *not* a perfect replacement for
@@ -33,7 +34,7 @@ pub trait EventLoopExtRunReturn {
     /// underlying OS APIs, which cannot be hidden by `winit` without severe stability repercussions.
     ///
     /// You are strongly encouraged to use `run`, unless the use of this is absolutely necessary.
-    fn run_return<F>(&mut self, event_handler: F) -> i32
+    fn run_return<F>(&mut self, event_handler: F) -> u8
     where
         F: FnMut(
             Event<'_, Self::UserEvent>,
@@ -45,7 +46,7 @@ pub trait EventLoopExtRunReturn {
 impl<T> EventLoopExtRunReturn for EventLoop<T> {
     type UserEvent = T;
 
-    fn run_return<F>(&mut self, event_handler: F) -> i32
+    fn run_return<F>(&mut self, event_handler: F) -> u8
     where
         F: FnMut(
             Event<'_, Self::UserEvent>,
